feat: add Placeholder and lua_State* support to LuaRegister
To support registration of a function that requies the lua state, it is
now possible to register a function that takes a lua_State*, which will
pass the same lua state as is used when calling LuaRegister::Register.

Placeholder can be used to ignore a parameter from lua. The assumtion is
that this parameter will be handled by the function that is being
registered. In the registered function, the parameter will correspond to
the lua stack index of the ignored parameter.

See the next commit for an example